telegram_bot/requirements.txt
python-telegram-bot==20.8
requests>=2.28.0
python-dotenv>=1.0.0
